MOve api routes to apiRouter
